complete t9-delete-comment. Created new 'comments' controller, moving old related code into it
