+++ __tests__/app.test.js
@@ -339,6 +339,37 @@ describe("/api/articles/:article_id/comments", () => {
     });
 });
 
+describe("/api/comments/:comment_id", () => {
+    test("DELETE204: successfully deletes comment by id, returning no content", () => {
+        return request(app)
+            .delete("/api/comments/5")
+            .expect(204)
+            .then(({ body }) => {
+                expect(body).toEqual({});
+                return db.query(`SELECT * FROM comments WHERE comment_id = 5`);
+            })
+            .then(({ rows }) => {
+                expect(rows).toHaveLength(0);
+            });
+    });
+    test("DELETE400: responds with appropriate error message when comment_id is not a valid id", () => {
+        return request(app)
+            .delete("/api/comments/not-valid-id")
+            .expect(400)
+            .then(({ body }) => {
+                expect(body.msg).toBe("invalid id");
+            });
+    });
+    test("DELETE404: responds with appropriate error message when comment_id is valid but doesn't exist", () => {
+        return request(app)
+            .delete("/api/comments/98327")
+            .expect(404)
+            .then(({ body }) => {
+                expect(body.msg).toBe("comment does not exist");
+            });
+    });
+});
+
 describe("Generic error handling", () => {
     test("GET404: should respond with a 404 error if endpoint doesn't exist", () => {
         return request(app).get("/api/no-endpoint").expect(404);
