+++ endpoints.json
@@ -95,5 +95,10 @@
                 "article_id": 7
             }
         }
+    },
+    "DELETE /api/comments/:comment_id": {
+        "description": "deletes a comment by comment_id. No response is returned.",
+        "queries": [],
+        "exampleResponse": {}
     }
 }
