+++ models/articles-model.js
@@ -51,34 +51,3 @@ exports.updateArticleById = (article_id, inc_votes) => {
             return rows[0];
         });
 };
-
-exports.selectCommentsByArticleId = (article_id) => {
-    return db
-        .query(
-            `SELECT comments.*
-             FROM comments
-             LEFT JOIN articles
-             ON comments.article_id = articles.article_id
-             WHERE comments.article_id = $1
-             ORDER BY created_at DESC`,
-            [article_id]
-        )
-        .then(({ rows }) => {
-            return rows;
-        });
-};
-
-exports.insertCommentByArticleId = (commentToPost) => {
-    return db
-        .query(
-            `INSERT INTO comments
-            (article_id, author, body)
-        VALUES
-            ($1, $2, $3)
-        RETURNING *`,
-            commentToPost
-        )
-        .then(({ rows }) => {
-            return rows[0];
-        });
-};
